requirements.txt
# Core data processing
pandas==2.0.3
numpy==1.24.3

# Machine learning
scikit-learn==1.3.1
xgboost==2.0.0
imbalanced-learn==0.11.0

# API
fastapi==0.104.1
uvicorn==0.24.0

# Configuration
pydantic==2.3.0
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.9.1
flake8==6.0.0
